EPMRPP-112010 || Bump ui-kit version and fix circular dependency log/storageUtils <-> user/sagas (#4908)
