+++ app/package.json
@@ -24,7 +24,7 @@
     "@formatjs/intl-pluralrules": "1.3.9",
     "@formatjs/intl-relativetimeformat": "4.5.1",
     "@formatjs/intl-utils": "1.6.0",
-    "@reportportal/ui-kit": "^0.0.1-alpha.173",
+    "@reportportal/ui-kit": "^0.0.1-alpha.174",
     "axios": "^1.13.2",
     "c3": "0.7.20",
     "chart.js": "2.9.4",
