+++ app/src/controllers/log/storageUtils.js
@@ -15,23 +15,15 @@
  */
 
 import { MARKDOWN } from 'common/constants/logViewModes';
-import { LOG_TIME_FORMAT_ABSOLUTE } from 'controllers/user/constants';
 import { getUserSettingsFromStorage, updateUserSettingsInStorage } from 'controllers/user';
 import {
   LOG_LEVEL_STORAGE_KEY,
   LOG_VIEW_MODE_STORAGE_KEY,
-  LOG_TIME_FORMAT_STORAGE_KEY,
 } from './constants';
 import { isLogLevelsEqual, isDefaultLogLevel } from './utils';
 
 const getLogLevelFromStorage = (userId) =>
   getUserSettingsFromStorage(userId)[LOG_LEVEL_STORAGE_KEY];
-
-export const getLogTimeFormatFromStorage = (userId) =>
-  getUserSettingsFromStorage(userId)[LOG_TIME_FORMAT_STORAGE_KEY] || LOG_TIME_FORMAT_ABSOLUTE;
-
-export const setLogTimeFormatInStorage = (userId, format) =>
-  updateUserSettingsInStorage(userId, { [LOG_TIME_FORMAT_STORAGE_KEY]: format });
 
 const getLogViewModeFromStorage = (userId) =>
   getUserSettingsFromStorage(userId)[LOG_VIEW_MODE_STORAGE_KEY];
