+++ app/src/controllers/user/index.js
@@ -29,6 +29,7 @@ export {
   LOGS_SIZE_KEY,
   LOGS_FULL_WIDTH_MODE_KEY,
   LOGS_COLORIZED_BACKGROUND_KEY,
+  LOG_TIME_FORMAT_STORAGE_KEY,
 } from './constants';
 export {
   fetchUserAction,
@@ -89,4 +90,6 @@ export {
   setLogsSizeInStorage,
   setLogsFullWidthModeInStorage,
   setLogsColorizedBackgroundInStorage,
+  getLogTimeFormatFromStorage,
+  setLogTimeFormatInStorage,
 } from './storageUtils';
